[ALPS04743553] sensors: fix converity issue
1. fix invalid type in argument ot print: atomic64_read
retruns long type but need long long, so fix to int64_t
2. fix dereference after null check: if device->scp_fifo
is null, it should report bug to kernel. if c->notifier_call
or monitor->name or monitor->notifier_call is null,
it should avoid the next operation

MTK-Commit-Id: cdf0dda16f6ad31de7fb3d247bd2902cbb64d303

CR-Id: ALPS04743553
Feature: Sensor Hub

+++ drivers/misc/mediatek/sensor/2.0/core/hf_manager.c
@@ -901,7 +901,7 @@ static int hf_manager_proc_show(struct seq_file *m, void *v)
 			continue;
 		seq_printf(m, "manager: param:[%d,%lld]\n",
 			atomic_read(&manager->io_enabled),
-			atomic64_read(&manager->io_poll_interval));
+			(int64_t)atomic64_read(&manager->io_poll_interval));
 		seq_printf(m, "device:%s poll:%s bus:%s online\n",
 			device->dev_name,
 			device->device_poll ? "io_polling" : "io_interrupt",
@@ -931,7 +931,8 @@ static int hf_manager_proc_show(struct seq_file *m, void *v)
 				client->request[i].enable,
 				client->request[i].delay,
 				client->request[i].latency,
-				atomic64_read(&client->request[i].start_time));
+				(int64_t)atomic64_read(
+					&client->request[i].start_time));
 		}
 	}
 	spin_unlock_irqrestore(&hf_client_list_lock, flags);
